pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prof-filter"
version = "0.1.0"
description = "Seven-school faculty aggregation MVP for public computer science faculty information"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.115.0",
  "httpx>=0.27.2",
  "jinja2>=3.1.4",
  "pydantic>=2.9.2",
  "pyyaml>=6.0.2",
  "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[tool.setuptools]
packages = ["app", "app.services", "crawler", "crawler.core", "crawler.spiders"]

[tool.pytest.ini_options]
testpaths = ["tests"]
